Warn if python version is unsupported
Issue: #71
